[TASK] Configure controller to use SomeaTemplateView
Due to a bug in FLOW3 1.0, it was possible to configure
property injection in Objects.yaml for a class which
was then also inherited to sub classes. The controllers
which inherited from ActionController directly relied
on this fact.

In FLOW3 1.1 this was fixed and therefore each controller
must either define the defaultViewObjectName itself or
inherit another controller which does that (AbstractModuleController).